requirements.txt
django
djangorestframework-simplejwt
mysqlclient
drf_yasg
djangorestframework
graphene-django
